third_party/blink/web_tests/http/tests/navigation/post-frames-goback1-expected.txt

Tests that a POST targetted to a frame is handled correctly when navigating away and coming back to the page.

--------
Frame: 'target-frame'
--------
This page was requested with the HTTP method POST.

Parameters:

the-input = input value goes here

Http headers:

HTTP_CACHE_CONTROL = max-age=0
HTTP_CONNECTION = keep-alive
HTTP_HOST = 127.0.0.1:8000
HTTP_ORIGIN = http://127.0.0.1:8000
HTTP_REFERER = http://127.0.0.1:8000/navigation/post-frames-goback1.html
HTTP_SEC_CH_UA = content_shell 99
HTTP_SEC_FETCH_DEST = iframe
HTTP_SEC_FETCH_MODE = navigate
HTTP_SEC_FETCH_SITE = same-origin
HTTP_UPGRADE_INSECURE_REQUESTS = 1

============== Back Forward List ==============
curr->  http://127.0.0.1:8000/navigation/post-frames-goback1.html
            http://127.0.0.1:8000/navigation/resources/form-target.pl (in frame "target-frame")
        http://127.0.0.1:8000/navigation/resources/go-back.html
===============================================
